Switch to dulwich
pygit2 is considerably faster but it's a pain to install because of how
it pins to particular versions of libgit and setting it up to have
multiple versions of libgit available is very annoying.

I initially used dulwich in harpoon before I used libgit, but the way I
was doing it was slow. By using the same method I was using with libgit
I'm able to get it down to acceptable speeds.

For example, on a repository with with files spanning 1500 commits, it
takes 3 seconds, a repository spanning 2500 commits takes 5 seconds.

This isn't as good as under one second with pygit2, but it's good enough